formatValidationErrors is now a method of Utils
